Fix day change bug
When the day changes, WeeChat prints the following message:
    Day changed to Thu, 30 Aug 2012
This causes the 'tags' list in the dictionary to be empty. Because we're
not interested in getting notified of day changes, the loop should pass.

Further, rather than using yaml.items() to get at the values, yaml.values()
is uesed to get to the data. Also make stylistic changes to the code, and
add comments.